requirement.txt
apscheduler>=3.3.0
aws-requests-auth>=0.3.0
blist>=1.3.6
configparser>=3.5.0
elasticsearch>=7.0.0
jsonschema>=3.0.2
PyStaticConfiguration>=0.10.3
python-dateutil>=2.6.0,<2.7.0
PyYAML>=5.1
requests>=2.0.0
